updated stability so that it is actually called and runs
