+++ bin/model_funcs/bcfar.py
@@ -10,8 +10,10 @@ def far_field(bcmodel, model, workspace, state):
     # set geometry parameters
     pad = model.padding
     [nx, ny] = workspace.field_size()
+    [nxp, nyp] = [pad +nx+pad, pad +ny+pad]
     [il, jl] = [nx+1, ny+1]
     [ie, je] = [nx+2, ny+2]
+    [ib, jb] = [nx+3, ny+3]
 
     # get physical parameters
     mp = model.params
